Re-use Petition component for preview, abstract Petitions actions
